clean up subscribe
