+++ app-shell/src/notifications/connect.ts
@@ -4,10 +4,10 @@ import { FAILURE_STATUSES, HEALTH_STATUS_OK } from '../constants'
 import { connectionStore } from './store'
 import { notifyLog } from './log'
 import { sendToBrowserDeserialized, deserialize } from './deserialize'
-import { unsubscribe } from './subscribe'
 
 import type { NotifyTopic } from '@opentrons/app/lib/redux/shell/types'
 import type { DiscoveryClientRobot } from '@opentrons/discovery-client'
+import { unsubscribe } from './unsubscribe'
 
 // MQTT is somewhat particular about the clientId format and will connect erratically if an unexpected string is supplied.
 // This clientId is derived from the mqttjs library.
