+++ app-shell/src/notifications/subscribe.ts
@@ -1,11 +1,14 @@
+import mqtt from 'mqtt'
+
+import { connectionStore } from './store'
+import { sendToBrowserDeserialized } from './deserialize'
+import { notifyLog } from './log'
+import { FAILURE_STATUSES } from '../constants'
+
 import type {
-  NotifyBrokerResponses,
   NotifyNetworkError,
   NotifyTopic,
 } from '@opentrons/app/lib/redux/shell/types'
-import { FAILURE_STATUSES } from '../constants'
-import mqtt from 'mqtt'
-import { connectionStore } from '../notify'
 
 /**
  * @property {number} qos: "Quality of Service", "at least once". Because we use React Query, which does not trigger
@@ -48,7 +51,7 @@ export function subscribe({
           } else {
             void waitUntilActiveOrErrored('subscription').catch(
               (error: Error) => {
-                log.debug(error.message)
+                notifyLog.debug(error.message)
                 sendToBrowserDeserialized({
                   hostname,
                   topic,
@@ -60,13 +63,55 @@ export function subscribe({
         }
       })
       .catch((error: Error) => {
-        log.debug(error.message)
+        notifyLog.debug(error.message)
         sendToBrowserDeserialized({
           hostname,
           topic,
           message: FAILURE_STATUSES.ECONNFAILED,
         })
       })
+  }
+
+  function subscribeCb(error: Error, result: mqtt.ISubscriptionGrant[]): void {
+    const { subscriptions } = connectionStore[hostname]
+    if (error != null) {
+      sendToBrowserDeserialized({
+        hostname,
+        topic,
+        message: FAILURE_STATUSES.ECONNFAILED,
+      })
+    } else {
+      notifyLog.debug(
+        `Successfully subscribed on ${hostname} to topic: ${topic}`
+      )
+      subscriptions.add(topic)
+    }
+  }
+  // Check every 500ms for 2 seconds before failing.
+  function waitUntilActiveOrErrored(
+    connection: 'client' | 'subscription'
+  ): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      const MAX_RETRIES = 4
+      let counter = 0
+      const intervalId = setInterval(() => {
+        const host = connectionStore[hostname]
+        const hasReceivedAck =
+          connection === 'client'
+            ? host.client != null
+            : host.subscriptions.has(topic)
+        if (hasReceivedAck) {
+          clearInterval(intervalId)
+          resolve()
+        }
+
+        counter++
+        if (counter === MAX_RETRIES) {
+          clearInterval(intervalId)
+          reject(new Error('Maximum number of retries exceeded.'))
+        }
+      }, CHECK_CONNECTION_INTERVAL)
+    })
   }
 }
 
@@ -80,77 +125,4 @@ function determineErrorMessageFor(hostname: string): NotifyNetworkError {
     message = FAILURE_STATUSES.ECONNFAILED
   }
   return message
-}
-
-function subscribeCb(error: Error, result: mqtt.ISubscriptionGrant[]): void {
-  const { subscriptions } = connectionStore[hostname]
-  if (error != null) {
-    sendToBrowserDeserialized({
-      hostname,
-      topic,
-      message: FAILURE_STATUSES.ECONNFAILED,
-    })
-  } else {
-    log.debug(`Successfully subscribed on ${hostname} to topic: ${topic}`)
-    subscriptions.add(topic)
-  }
-}
-
-// Check every 500ms for 2 seconds before failing.
-function waitUntilActiveOrErrored(
-  connection: 'client' | 'subscription'
-): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    const MAX_RETRIES = 4
-    let counter = 0
-    const intervalId = setInterval(() => {
-      const host = connectionStore[hostname]
-      const hasReceivedAck =
-        connection === 'client'
-          ? host.client != null
-          : host.subscriptions.has(topic)
-      if (hasReceivedAck) {
-        clearInterval(intervalId)
-        resolve()
-      }
-
-      counter++
-      if (counter === MAX_RETRIES) {
-        clearInterval(intervalId)
-        reject(new Error('Maximum number of retries exceeded.'))
-      }
-    }, CHECK_CONNECTION_INTERVAL)
-  })
-}
-
-export function unsubscribe(
-  hostname: string,
-  topic: NotifyTopic
-): Promise<void> {
-  return new Promise<void>(() => {
-    if (!pendingUnsubs.has(topic)) {
-      if (connectionStore[hostname].subscriptions.has(topic)) {
-        pendingUnsubs.add(topic)
-        const { client } = connectionStore[hostname]
-        client?.unsubscribe(topic, {}, (error, result) => {
-          if (error != null) {
-            log.debug(
-              `Failed to unsubscribe on ${hostname} from topic: ${topic}`
-            )
-          } else {
-            log.debug(
-              `Successfully unsubscribed on ${hostname} from topic: ${topic}`
-            )
-            const { subscriptions } = connectionStore[hostname]
-            subscriptions.delete(topic)
-            pendingUnsubs.delete(topic)
-          }
-        })
-      } else {
-        log.debug(
-          `Host ${hostname} to unsubscribe from unsubscribed topic: ${topic}`
-        )
-      }
-    }
-  })
 }
